[OV15] Better implementation of possessed summons.
New subclass of creature (similar to TemporarySummons) that is fully controlled
(movement, camera...) by the player on summon. We mostly improve the
bound-to-aura character of possesed summons.

I must admit, that the handling in the SpellAuraHolder is not a very
pretty one, but at least it work's at it should!